backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
python-dotenv==1.2.2
pydantic==2.12.5
pydantic-settings==2.13.1
llama-index-core==0.14.15
llama-index-readers-file==0.5.6
llama-index-vector-stores-chroma==0.5.5
llama-index-embeddings-openai==0.5.2
llama-index-llms-openai==0.6.26
chromadb==1.5.2
pymupdf==1.27.1
tiktoken==0.12.0
pytest>=8.0
httpx>=0.27
